pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pdfplumber>=0.11.7",
    "python-docx>=1.2.0",
    "reportlab>=4.4.2",
    "streamlit>=1.46.1",
]
